pyproject.toml
[project]
name = "vigil"
version = "0.1.0"
description = "Real-Time Traffic Accident Detection System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "debugpy>=1.8.8",
    "ffmpeg-python>=0.2.0",
    "fire>=0.7.0",
    "inference>=0.25.1",
    "influxdb-client>=1.48.0",
    "jsonpickle>=3.3.0",
    "python-dotenv>=1.0.1",
    "zmq>=0.0.0",
]

[tool.pytest.ini_options]
pythonpath = "."
